refactor(test/controller): 컨트롤러 수정에 의한 테스트 코드 수정
